Implement Silicon Labs EFR32 Wi-SUN support (#22)
* Add osal/freertos_efr32

* Add EFR32 Wi-SUN platform dir content to the repository

* Implement efr32 based csmp agent lib build

* Change random() function to arm-gcc stdlib rand()

* Implement sample efr32_wisun.target makefile

* Replace 'settimeofday()' with osal_settime

* Fix linker file reference and clean 'efr32_wisun/efr32_wisun.mak'

* Add 'build/' to .gitignore

* Fix make files and build.sh to handle 'efr32_wisun' target properly

* Refactor linux.target makefile for lib build

* Refactor linux.target makefile for sample application

* Refactor freertos.target makefile for lib build

* Refactor freertos.target makefile for sample app

* Fix variable initialization warnings (droped by arm-gcc)

* Implement multi platform support for printf and scanf format specifiers

* Add '-Wno-missing-braces' to CFLAGS to supress warnings in 'CsmpTlvs.pb.c/h' files

* Implement osal_task_fnc_t and refactor thread functions accordingly

* Fix xTimerPendFunction type conversion for freertos and efr32_wisun targets

* Remove select from coapclient

(cherry picked from commit 3d213cdabcaf1798f8ae1c93f84f0dc783bcf743)

* Remove select from coapserver

(cherry picked from commit 8ce64ca83eb635b66a3c423e9529ba242494c369)

* Remove commented sections from efr32 wisun osal port

* Enable SL_SLEEPTIMER_WALLCLOCK_CONFIG' to have time setter/getter

* Implement get and settime using sl_sleeptimer API

* Implement compile time paramaters for FreeRTOS thread functions in sample app

* Implement CsmpAgentLib_sample_config.h to provide config interface

* Add rail lib and nvm3 lib exceptions to .gitignore

* Add system init call for efr32-wisun kernel start sequence

* Fix get and set time in osal efr32-wisun

* Add wisun connect and wait to app task

* Implement portable tlvs capability for sample application

* Change wpan and hw desc tlvs for efr32_wisun_tlvs

* Implement ip address get tlv for efr32_wisun

* Implement NTP client for EFR32 target to obtain time

* Update README.md with EFR32 wisun content

* Fix Linux target after rebase

* Fix FreeRTOS target after rebase

* Fix EFR32 Wi-SUN target after rebase

* Update gsdk to 'simplicity_sdk_2024.6.0'

* Eliminate linker warnings for freertos and linux targets

* Remove SiSDK dependencies and create Vendor directory

* Fix sample/efr32_wisun.target for supporting Windows build

* Add Vendor/Silabs generated content to .gitignore

* Update readme content with efr32 platform

* add support of the skeleton application to the readme

* Removing Wi-SUN SoC Csmp Agent Sample references from 'Vendors/Silabs/Readme.md'

* Remove wisun logo from Vendors/Silabs/Readme.md

* Remove CsmpAgentLib_sample_util source and header files

* Remove CsmpAgentLib_sample_config.h and put the content into CsmpAgentLib_sample.h

* Extend requirement list with exact versions in 'Vendors/Silabs/Readme.md'

* Add additional link to Silabs EFR32 products in 'README.md'

* Add Wi-SUN alliance standard logo to 'Vendors/Silabs/Readme.md'

* Remove unused resources from 'Vendors/Silabs/Readme.md'

* Fix typos in the readme files

* Revert periodic metrics timer changes

---------

+++ build.sh
@@ -24,8 +24,10 @@ build_sample()
 build_lib()
 {
   make $OS -C $DESTDIR
-  rm *.o
-  mv csmp_agent_lib.a sample/
+  rm -f *.o
+  [ ! -f csmp_agent_lib.a ] || mv csmp_agent_lib.a sample/
+  [ ! -f csmp_agent_lib_freertos.a ] || mv csmp_agent_lib_freertos.a sample/
+  [ ! -f csmp_agent_lib_efr32_wisun.a ] || mv csmp_agent_lib_efr32_wisun.a sample/
 }
 
 clean_all()
@@ -33,6 +35,9 @@ clean_all()
   make clean -C $DESTDIR
   make clean -C $SAMPLEDIR
 #  make clean -C $TLVPROTODIR
+  rm -rf $DESTDIR/build
+  rm -f $DESTDIR/*.a
+  rm -f $SAMPLEDIR/*.a
 }
 
 if [ "$1"x = "clean"x ];then
